requirements.txt
anthropic>=0.40.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.0
